2025/June-11/calculator-demo/requirements.txt
python-dotenv==1.0.0
openai==1.84.0
requests==2.31.0
langchain-openai
langchain
